+++ staffflow-server/src/main/java/kz/zzhalelov/staffflow/server/dto/departmentDto/DepartmentUpdateDto.java
@@ -8,5 +8,5 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class DepartmentUpdateDto {
     String name;
-    Integer managerId;
+    Long managerId;
 }
