+++ staffflow-server/src/main/java/kz/zzhalelov/staffflow/server/repository/LaborContractRepository.java
@@ -3,5 +3,5 @@ package kz.zzhalelov.staffflow.server.repository;
 import kz.zzhalelov.staffflow.server.model.LaborContract;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LaborContractRepository extends JpaRepository<LaborContract, Integer> {
+public interface LaborContractRepository extends JpaRepository<LaborContract, Long> {
 }
